+++ src/Oro/Bundle/ConfigBundle/Tests/Unit/Config/UserConfigManagerTest.php
@@ -42,8 +42,8 @@ class UserConfigManagerTest extends \PHPUnit_Framework_TestCase
         $this->object = new UserConfigManager($this->om, $this->settings);
 
         $this->tokenStorage = $this->createMock('Symfony\Component\Security\Core\TokenStorageInterface');
-        $this->group1 = $this->createMock('Oro\Bundle\UserBundle\Entity\Group');
-        $this->group2 = $this->createMock('Oro\Bundle\UserBundle\Entity\Group');
+        $this->group1 = $this->createMock('Pim\Component\User\Model\Group');
+        $this->group2 = $this->createMock('Pim\Component\User\Model\Group');
 
         $token = $this->createMock('Symfony\Component\Security\Core\Authentication\Token\TokenInterface');
         $user = new User();
