+++ src/Oro/Bundle/UserBundle/Controller/GroupController.php
@@ -3,8 +3,8 @@
 namespace Oro\Bundle\UserBundle\Controller;
 
 use Oro\Bundle\SecurityBundle\Annotation\AclAncestor;
-use Oro\Bundle\UserBundle\Entity\Group;
 use Oro\Bundle\UserBundle\OroUserEvents;
+use Pim\Component\User\Model\Group;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\EventDispatcher\GenericEvent;
