+++ src/Oro/Bundle/UserBundle/Form/Type/GroupType.php
@@ -2,8 +2,8 @@
 
 namespace Oro\Bundle\UserBundle\Form\Type;
 
-use Oro\Bundle\UserBundle\Entity\Group;
 use Pim\Bundle\EnrichBundle\Form\Type\EntityIdentifierType;
+use Pim\Component\User\Model\Group;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
