+++ src/Oro/Bundle/UserBundle/Tests/Unit/Entity/GroupTest.php
@@ -3,7 +3,7 @@
 namespace Oro\Bundle\UserBundle\Tests\Entity;
 
 use Doctrine\Common\Collections\ArrayCollection;
-use Oro\Bundle\UserBundle\Entity\Group;
+use Pim\Component\User\Model\Group;
 use Pim\Component\User\Model\Role;
 
 class GroupTest extends \PHPUnit_Framework_TestCase
