+++ src/Pim/Component/User/Model/Group.php
@@ -1,14 +1,10 @@
 <?php
 
-namespace Oro\Bundle\UserBundle\Entity;
+namespace Pim\Component\User\Model;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
-use JMS\Serializer\Annotation\Exclude;
-use JMS\Serializer\Annotation\Type;
-use Pim\Component\User\Model\GroupInterface;
-use Pim\Component\User\Model\Role;
 
 /**
  * @ORM\Entity(repositoryClass="Oro\Bundle\UserBundle\Entity\Repository\GroupRepository")
